feature:add python encoder wrapper
